+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * Updates to readme
 * Updated React 15.2.1 to version 15.3.1
 * Updated React Native 0.31.0 to version 0.32.0
+* Updated PropTypes declaration to correct array
 
 ### 0.6.0 (August 19, 2016)
 * Forked from react-native-popup